+++ src/common/sub-interface/spin-text/st-img-opts.js
@@ -1,0 +1,11 @@
+const imgOptProgString = "Validating Image Options";
+const imgOptCompString = "Image Options Valid";
+const imgOptFailString = "Invalid Image Options";
+
+
+module.exports =
+{
+	imgOptProg: imgOptProgString,
+	imgOptComp: imgOptCompString,
+	imgOptFail: imgOptFailString
+};