+++ src/input/image-options-validation.js
@@ -2,27 +2,48 @@ const ora = require("ora");
 const configHelp = require("./image-components/config-help");
 const hexColour = require("./arguments/image-options/hex-colour");
 const imageNumbers = require("./arguments/image-options/image-numbers");
+const spinText = require("../common/sub-interface/spin-text/st-img-opts");
 
 
 // Validates image parsing options.
 function prepareImageOptionArguments(fullInputObject, optionalArgs, imageOptionCallback)
 {
-	var preperationSpinner = ora("Validating Image Options").start();
+	var preperationSpinner = ora(spinText.imgOptProg).start();
 	
 	coordinateOptionValidation(fullInputObject, optionalArgs, function(argError, argRes)
 	{
 		if (argError !== null)
 		{
-			preperationSpinner.fail("Invalid Image Options");
+			preperationSpinner.fail(spinText.imgOptFail);
 			return imageOptionCallback(argError, null);
 		}
 		else
 		{
-			preperationSpinner.succeed("Image Options Valid");
+			preperationSpinner.succeed(spinText.imgOptComp);
 			return imageOptionCallback(null, true);
 		}
 	});
 	
+}
+
+
+function prepareImageOutputArguments(fullInputObject, optionalArgs, imageOutputCallback)
+{
+	var preperationSpinner = ora(spinText.imgOptProg).start();
+	
+	coordinateOutputValidation(fullInputObject, optionalArgs, function(argError, argRes)
+	{
+		if (argError !== null)
+		{
+			preperationSpinner.fail(spinText.imgOptFail);
+			return imageOutputCallback(argError, null);
+		}
+		else
+		{
+			preperationSpinner.succeed(spinText.imgOptComp);
+			return imageOutputCallback(null, true);
+		}
+	});
 }
 
 
@@ -59,12 +80,12 @@ function coordinateOptionValidation(fullInput, optArgs, imgOptCallback)
 	
 	
 	// Begin option validation:
-	wallColourValid = hexColour.validateHexColour(givenWallColour, "wallColour", "Wall Colour", fullInput, validationOutcomeObject);
+	wallColourValid = handleWallValidation(givenWallColour, fullInput, validationOutcomeObject);
 	
 	
 	if (wallColourValid === true)
 	{
-		floorColourValid = hexColour.validateHexColour(givenFloorColour, "floorColour", "Floor Colour", fullInput, validationOutcomeObject);
+		floorColourValid = handleFloorValidation(givenFloorColour, fullInput, validationOutcomeObject);
 	}
 	
 	if (floorColourValid === true)
@@ -93,8 +114,6 @@ function coordinateOptionValidation(fullInput, optArgs, imgOptCallback)
 	}
 	
 	
-	
-	
 	if (validationOutcomeObject.overallValid === true)
 	{
 		return imgOptCallback(null, true);
@@ -108,8 +127,72 @@ function coordinateOptionValidation(fullInput, optArgs, imgOptCallback)
 
 
 
+function coordinateOutputValidation(fullInput, optArgs, argCallback)
+{
+	var validationOutcomeObject = configHelp.defineValidationResult();
+	
+	var givenWallColour = null;
+	var givenFloorColour = null;
+	var givenTileSize = null;
+	
+	var wallColourValid = false;
+	var floorColourValid = false;
+	var toleranceValid = false;
+	var tileSizeValid = false;
+	
+	configHelp.overrideInternalProperties(fullInput.imageItems, optArgs);
+	
+	givenWallColour = fullInput.imageItems.wallColour;
+	givenFloorColour = fullInput.imageItems.floorColour;
+	givenTileSize = fullInput.imageItems.tileSize;
+	
+	wallColourValid = handleWallValidation(givenWallColour, fullInput, validationOutcomeObject);
+	
+	if (wallColourValid === true)
+	{
+		floorColourValid = handleFloorValidation(givenFloorColour, fullInput, validationOutcomeObject);
+	}
+	
+	if (floorColourValid === true)
+	{
+		tileSizeValid = imageNumbers.validateTileSize(givenTileSize, fullInput, validationOutcomeObject);
+	}
+	
+	if (tileSizeValid === true)
+	{
+		validationOutcomeObject.overallValid = true;
+	}
+	
+	if (validationOutcomeObject.overallValid === true)
+	{
+		return argCallback(null, true);
+	}
+	else
+	{
+		return argCallback(new Error(validationOutcomeObject.errorMessage), null);
+	}
+	
+}
+
+
+function handleWallValidation(inpValue, fullInpObj, validOutcomeObj)
+{
+	var handleRes = hexColour.validateHexColour(inpValue, "wallColour", "Wall Colour", fullInpObj, validOutcomeObj);
+	return handleRes;
+}
+
+
+function handleFloorValidation(inpValue, fullInpObj, validOutcomeObj)
+{
+	var handleRes = hexColour.validateHexColour(inpValue, "floorColour", "Floor Colour", fullInpObj, validOutcomeObj);
+	return handleRes;
+}
+
+
+
 
 module.exports =
 {
-	prepareOptionArguments: prepareImageOptionArguments
+	prepareOptionArguments: prepareImageOptionArguments,
+	prepareOutputArguments: prepareImageOutputArguments
 };